Updated tools file. Merge did not work. Check why it failed.

+++ cellorganizer/tools.py
@@ -187,14 +187,18 @@ def slml2report(model1_filename, model2_filename, options):
 
     f = open(txtfilename,"a")
     f.write("model1_filename =" + model1_filename + "\n")
-    f.write("model1_filename =" + model2_filename + "\n")
+    f.write("model2_filename =" + model2_filename + "\n")
+    f.close()
 
-    os.system("slml2report input.txt;rm input")
+    os.system("slml2report input.txt;rm input.txt")
 
     # need to copy everything out of report folder to current working directory
-    os.system('mv ./report/* .')
-    os.system('rm -r report/')
-    return None
+    if os.path.exists('report'):
+        os.system('mv ./report/* .')
+        os.system('rm -rf ./report')
+        return True
+    else:
+        return False
 
 #######################################################################
 def imshow(img_path, options):
